Add files via upload
A Python script for a client in the agriculture industry to extract current price updates for vegetables, fruits, and ornamentals from their website. The data was then subjected to exploratory data analysis (EDA) and machine learning algorithms to identify patterns and predict future price trends. Results were inserted into a PostgreSQL database and displayed in a real-time dashboard for the client. The project followed best practices in data science, including modular programming, data cleaning, and normalization.